msm: ipa_qmi: handle is good in ipa3_a5_svc_req_cb()
There is no need to verify that the handle provided to
ipa3_a5_svc_req_cb()

That function is only called as a req_cb function for
ipa3_a5_svc_ops_options.  The req_cb function is only called
by handle_qmi_request().  And the only reason that function would
call this particular request function is if it was associated with
the handle that's supplied to it.  So the handle supplied to
ipa3_a5_svc_req_cb() *will* be ipa3_svc_handle.

Change-Id: I893e78573573c65051ac6dc3f7be08182f05502d
Signed-off-by: Alex Elder <elder@codeaurora.org>