msm: ipa_qmi: consistenly format offset field values
Define the value of offset field of each element info array entry
with consistent formatting and indentation.  In a few cases the line
length exceeds 80 columns, but the real problem is the length of the
type/field name and I'm not fixing that here.  Like the previous
patch, this only affects formatting.

Change-Id: I1dab4bb3207ec6a3e5f35d3e177da31988e4fd02
Signed-off-by: Alex Elder <elder@codeaurora.org>