msm: ipa_qmi: consistenly format elem_size field values
Define the value of elem_size field of each element info array entry
with consistent formatting and indentation.  In a one case the line
length exceeds 80 columns, but the real problem is the length of the
type/field name and I'm not fixing that here.  This patch only
affects formatting.

Change-Id: I8c558f2c3fbfab5d526b6b2bbaf3a03976359a58
Signed-off-by: Alex Elder <elder@codeaurora.org>