msm: ipa_qmi: introduce ipa3_handle_config_req()
If ipa3_a5_svc_req_cb() receives a QMI_IPA_CONFIG_REQ_V01 request
from the modem, it ignores it and returns 0.  The caller is
handle_qmi_request(), which calls it as the QMI req_cb handler.
This is a problem because handle_qmi_request() will only return a
response to the modem if the req_cb function returns an error, so
the modem will never receive the response message it expects.

A CONFIG request response message contains only a "resp" field to
hold a result code is stored.  Define ipa3_handle_config_req() to,
which return a success response for CONFIG requests.

Change-Id: I62cae2d3058107f61805ed1eb8cff6d77c5c0c12
Signed-off-by: Alex Elder <elder@codeaurora.org>