msm: ipa_qmi: introduce ipa3_install_filter_rule_req()
This patch is very much like the previous one.

If ipa3_a5_svc_req_cb() receives a QMI_IPA_INSTALL_FILTER_RULE_REQ_V01
request from the modem, it ignores it and returns 0.  The caller is
handle_qmi_request(), which calls it as the QMI req_cb handler.
Since handle_qmi_request() will only return a response to the modem
if the req_cb function returns an error, the modem will never
receive the response message it expects.

Only the "resp" field of a INSTALL_FILTER_RULE response message
is mandatory, so we don't need to fill in all its fields.  Define
ipa3_install_filter_rule_req() to send a success response to
INSTALL_FILTER_RULE requests from the modem.

Change-Id: Ib0cdbfa310c68c1888a20a20d740a3d1a8b12b19
Signed-off-by: Alex Elder <elder@codeaurora.org>