msm: ipa_uc: use local variables in ipa3_uc_send_cmd()
Define a local variable "uc_ctx" to hold the address of the uc_ctx
field in the IPA context, and use it to make the code a little
easier to read.

Define another local variable "mmio" to hold the address of the
memory-mapped microcontroller memory region, to aid readability.
Reorder the assignments so the mmio fields are done in the order
they are defined in the IpaHwSharedMemCommonMapping_t structure.
Group the IPA uc_ctx assignments separate from the mmio ones.

Change-Id: Ibf63dda9e206758ac6a67bb996b62fdd3d6658a3
Signed-off-by: Alex Elder <elder@codeaurora.org>