msm: ipa_uc: introduce uc_command_should_retry()
Create a simple helper function to tell whether a microcontroller
command status indicates the command should be retried.  Use it
in _ipa3_uc_send_cmd() to determine if a command should retry,
and of so return -EAGAIN.

Have ipa3_uc_send_cmd() return immediately if a command completed
with an error that should not be retried, otherwise increment the
retry count.  Change the logic for retrying commands from two
consecutive "if" statements (whose taken branch ends with a "goto")
to a single "if...else" to make it obvious all cases are covered.

Change-Id: Ia48f7d744d3f91ee7230aa139b22fbd8c81835a2
Signed-off-by: Alex Elder <elder@codeaurora.org>