msm: ipa_uc: use common usleep ranges
The time ranges to sleep between retries for microcontroller
commands is 1-2 seconds regardless of the (retryable) command.
Define common symbols inside "ipa_uc.c" to define that range and use
them for both types of retried command.  For the channel empty case,
this eliminates the only use of the previous symbols so those can be
deleted.  For the channel stop case, the symbol is still used
elsewhere--but in that case it's direct GSI commands that are being
retried rather than microcontroller commands.  (I assert that's a
different situation, worthy of a separately-defined delay.)

Change-Id: Ic89b1d26b99def9b5996b60781753053ad38e14f
Signed-off-by: Alex Elder <elder@codeaurora.org>