msm: ipa_uc: use common return path in ipa3_uc_send_cmd()
We release the mutex after we're done sending a microcontroller
command, whether the command is successful or not (including
unrecoverable errors).  A timeout is not recoverable, as is
receiving too many producer disable command stop failures.  Any
non-recoverable error results in BUG() being called, which will
never return.

Rather than having mutex unlocks for every successful and
recoverable error return, use a single block at the end of the
function.  Use a return value of -EIO to signal that we've hit an
unrecoverable error.  By doing this we can have a single return path
for all cases.

Change-Id: Id8779dbe0e2c3b05674aac5d0b4b0cbe04c37c0b
Signed-off-by: Alex Elder <elder@codeaurora.org>