msm: ipa_uc: introduce try_send_uc_command()
Define a new command try_send_uc_command() to encapsulate trying to
send a command and interpreting a few possible status responses.
If it returns -EAGAIN, the command can be retried.  All other
returns can be handled in the normal return path at the end of
ipa3_uc_send_cmd().

Change-Id: Ib33563b26fa93fdb642819ee3d00924f2f209875
Signed-off-by: Alex Elder <elder@codeaurora.org>