msm: ipa_uc: private_data is good in ipa3_uc_response_hdlr()
There's no need to verify the private_data value provided to
is the same as ipa3_ctx in ipa3_uc_response_hdlr().  The only
way ipa3_uc_response_hdlr() will be called is as an interrupt
handler added by a call to ipa3_add_interrupt_handler()  in
ipa3_uc_interface_init().  And it will be called only from
ipa3_deferred_interrupt_work(), which will pass exactly the
private_data value that was supplied when the interrupt was
added.

Change-Id: I6b8d0256d7aab5c5be0ad81289a61396c4d267c2
Signed-off-by: Alex Elder <elder@codeaurora.org>