msm: ipa_uc: private_data is good in ipa3_uc_event_handler()
There's no need to verify the private_data value provided to
is the same as ipa3_ctx in ipa3_uc_event_handler().  The only
way ipa3_uc_event_handler() will be called is as an interrupt
handler added by a call to ipa3_add_interrupt_handler() in
ipa3_uc_interface_init().  And it will be called only from
ipa3_deferred_interrupt_work(), which will pass exactly the
private_data value that was supplied when the interrupt was
added.

Change-Id: I8adc81f6dc2e449f80f24ead99dfd26bdaf39e24
Signed-off-by: Alex Elder <elder@codeaurora.org>