msm: ipa_uc: take a shortcut in ipa3_log_evt_hdlr()
Check for the eventParams matching the the value we assigned it to
previously right away in ipa3_log_evt_hdlr(), and if they're equal,
just return.

This assumes the eventParams value will never be 0, but that was
already an implicit assumption.  If it were zero we would have
stashed that value in the uc_event_top_ofst field of the
microcontroller context.  But that being zero is a signal (in this
function) that it has never been assigned--and leads to
uc_ctx->uc_event_top_mmio being mapped (with no effort to release a
previously-mapped value).

Change-Id: I26f9ab73c979580be824a22664abdc9faa33e2f2
Signed-off-by: Alex Elder <elder@codeaurora.org>