msm: ipa_uc: check for changed value earlier in ipa3_log_evt_hdlr()
Again assuming that the eventParams value will never be zero, we can
check for a value that changed after we originally assigned it
easily near the top of ipa3_log_evt_hdlr(), and return immediately
after reporting that.

That allows the remainder of the function to be shifted left.  It
also makes it more natural to zero uc_event_top_ofst if the ioremap()
fails.

Change-Id: I7dcb221a15f7f984c0b12723913a56d52ebfdd00
Signed-off-by: Alex Elder <elder@codeaurora.org>