msm: ipa_dp: improve computation in ipa_adjust_ra_buff_base_sz()
The repeated OR operations in ipa_adjust_ra_buff_base_sz() produce
what the comment says it does--the highest power of two less than
the given size (after adding IPA_MTU and IPA_GENERIC_RX_BUFF_LIMIT,
that is).  But it's not very intuitive.

Instead, use the "find leading set" bit operation, to determine the
power of 2 and use that as a shift to produce the desired value.
Add a comment explaining why this works.

This has the side benefit of being easily optimized by the compiler
(finding the bit position is typically a single instruction).

Change-Id: Ic378881d28cd3ba3c0c1654e2e3217bd83849eea
Signed-off-by: Alex Elder <elder@codeaurora.org>