msm: ipa_dp: have ipa_poll_gsi_pkt() return size
There are two callers of have ipa_poll_gsi_pkt().  They do not check
for errors, they check for whether the returned value is 0 or not.

Rather than returning 0 or error, have ipa_poll_gsi_pkt() return the
size of the received block of data, or a negative errno on error.
The maximum size returned is limited by the 16 bit bytes_xferred
field in the gsi_chan_xfer_notify structure, so an int can represent
all sizes *and* negative error values.

Have the two callers check for a negative return instead of non-zero.

Change-Id: If4b7b81f7d369e27e2fbd868d4192ba9fc51f249
Signed-off-by: Alex Elder <elder@codeaurora.org>