msm: ipa_dp: notify is non-null in ipa_gsi_irq_rx_notify_cb()
The only way ipa_gsi_irq_rx_notify_cb() will ever be called is as
the xfer_cb function field in a GSI channel properties structure.
That function is called from gsi_process_chan(), which already
assumes the notify pointer is non-null (it dereferences it) so
it will never be null when ipa_gsi_irq_rx_notify_cb() is called.

Change-Id: I782c4a3dda4aee0cea3c18f75ef39e5f59e53c1a
Signed-off-by: Alex Elder <elder@codeaurora.org>