msm: ipa_qmi: simplify ipa3_q6_clnt_svc_event_notify()
This is less warranted than the previous one, but there are only two
evt_id values supported in ipa_gsi_irq_rx_notify_cb().  Replace the
switch statement with a single if statement that returns early if
the event id isn't recognized.  This allows most of the function to
be shifted left to improve readability.

Change-Id: I11d384a1d891cdaffacaa366514d82aa9ac9688e
Signed-off-by: Alex Elder <elder@codeaurora.org>