msm: ipa_utils: use ep_configuration() in ipa_init_ep_flt_bitmap()
Use ep_configuration() to get the configuration pointer for the
given client--moving toward only using that function when that is
needed.  The only reason it returns NULL is if the client provided
is out of range, but we know it is not (because of the range covered
by the loop).

Change-Id: Iea3538dbd9e4b3300b19163baa622f2961ad6a10
Signed-off-by: Alex Elder <elder@codeaurora.org>