msm: gsi: define GSI_EVT_RING_ELEMENT_SIZE
Define GSI_EVT_RING_ELEMENT_SIZE in "gsi.c" to be 16 bytes, and use
it wherever GSI_EVT_RING_RE_SIZE_16B was previously used.  This
removes the last use of the gsi_evt_ring_elem_size enumerated type
in the GSI lite code, but we'll leave the definition alone because
it's found in a file under /usr/include.

Change-Id: Ia194417b3abc4eb93d67b7cd2dc47aeb0ecfeec7
Signed-off-by: Alex Elder <elder@codeaurora.org>