msm: gsi: define GSI_CHAN_RING_ELEMENT_SIZE
Define GSI_CHAN_RING_ELEMENT_SIZE in "gsi.c" to be 16 bytes, and use
it wherever GSI_CHAN_RE_SIZE_16B was previously used.  This
removes the last use of the gsi_evt_ring_elem_size enumerated type
in the GSI lite code, but we'll leave the definition alone because
it's found in a file under /usr/include.

Change-Id: I073fd388a198209884d651cc6c4a0034d7585a78
Signed-off-by: Alex Elder <elder@codeaurora.org>