msm: ipahal: gsi_ctx is non-null everywhere.
The "gsi_ctx" global is initialized by the top-level IPA driver
probe function.  If it fails to be allocated, the driver probe fails
and none of the child drivers (specifically AP SMMU) will ever be
probed or loaded.  No GSI channel activities can occur until after gsi_register_device() is called in ipa3_post_init().

The only place we need to check for a non-null "gsi_ctx" pointer is
when it is allocated.  Delete these checks everywhere else they are
found in the GSI code.

Note:  It's conceivable a non-null "gsi_ctx" was assumed to mean
other things about the GSI state, but from what I can tell that's
not the case.

Change-Id: I456695581f4e4ebe43ef2b3bbfce3bed551e3df3
Signed-off-by: Alex Elder <elder@codeaurora.org>