msm: gsi: kill gsi_evt_ring_props->user_data
The user_data field of an gsi_evt_ring_props structure is
initialized before a call to gsi_alloc_evt_ring().  Both callers of
that function supply NULL for that, so it is really unused in the
IPA lite code.  Get rid of that field, and use NULL in the one place
that would used it, in gsi_handle_glob_err().

Change-Id: I39ca8c4214e6dac2d7d17cfd66f1f8f88f91fbb3
Signed-off-by: Alex Elder <elder@codeaurora.org>