msm: gsi: remove low-level debugging feature
GSIDBG_LOW() is called in one spot.  Temporarily comment that spot
out, and remove the definition of GSIDBG_LOW().  Remove support for
enabling this debugging for GSI via /sys/kernel/debug/gsi/ipc_low
by removing that debugfs file.  As a result, the ipc_logbuf_low
field of the GSI context will always be NULL, so GSI_IPC_LOGGING()
calls using that will be no-ops.  So they can be removed.

Change-Id: Ieeeeaf74c1a5981f76b2c8529c11599b343ecfac
Signed-off-by: Alex Elder <elder@codeaurora.org>