msm: gsi: use IPA log buffer
Have the GSI code share the IPA log buffer rather than creating its
own.  To do this, pass the log buffer handle from the IPA code to
the GSI code as an argument go msm_gsi_init().  Increase the size of
the IPA log buffer to account for the additional size previously
used for the GSI log buffer.

Change-Id: Idfaa2325ccff2fbc6c046960c922bf41b119628b
Signed-off-by: Alex Elder <elder@codeaurora.org>