msm: gsi: fix GSI_IPC_LOGGING()
The GSI_IPC_LOGGING() macro adds __func__ and __LINE__ to the set of
arguments passed to ipc_log_string(), which means the caller must be
sure to supply a "%s" and "%d" in the format string it provides.

Fix this by having these parameters in the same place as their
specifiers in the format string.

Change-Id: I69854bc6f83c5f0591ee7b30163c23ed7a06fe64
Signed-off-by: Alex Elder <elder@codeaurora.org>