msm: gsi: make ring context base_va a void pointer
The base_va field of a gsi_ring_ctx structure is defined to have
type unsigned long.  I'm not sure why this is--perhaps to ensure
arithmetic performed with that value is well-defined.  But it
represents a virtual address *pointer* and so should be defined
that way.  Making it a void pointer removes the need for a bunch
of casting.

One might argue that pointer arithmetic on a void pointer is not
well-defined--and for standard C it is not.  GCC treats void
pointer arithmetic as if a sizeof(void) = 1, meaning adding
to a void pointer "works the same" as adding to a char pointer.

If upstream reviewers balk at this usage I'll change it, but
it's a practice found elsewhere in the kernel code.

Change-Id: I45483dc431fa034508dbc70a3c708a0632d96c96
Signed-off-by: Alex Elder <elder@codeaurora.org>