msm: gsi: make gsi_ring_ctx->base a dma_addr_t
The base address in a ring context ultimately comes back from a DMA
allocation request, and is as such really represents a dma_addr_t.
Change the type of the "base" field, as well as a local variable
used in gsi_validate_evt_ring_props() that works with it.

Also, in preparation for the next patch, change the type of the
"len" field of a ring context to be u32.

Change-Id: I41a8e90051f555ebe5eac94d70d18a635bcc6322
Signed-off-by: Alex Elder <elder@codeaurora.org>