msm: ipa_dp: use ipahal_dma_alloc() for event ring memory
Now that the event ring memory information is stored in
an ipa_mem_info structure we can use ipahal_dma_alloc()
to allocate it, and ipahal_dma_free() to release it.

Note that in ipa_gsi_setup_channel() the error path will
only be taken if the DMA allocation is successful, so
there's no need to check whether the address is non-null
before freeing.

Change-Id: Ifcc9c3ebd377b0543a52ec2ea8dc7a99b77cd47b
Signed-off-by: Alex Elder <elder@codeaurora.org>