msm: gsi: have gsi_alloc_evt_ring() allocate memory
There are two callers of gsi_alloc_evt_ring(), and each is supposed
to supply an allocated range of DMA memory in the event ring
properties they provide.  The only information needed is the size of
the space to allocate; both use GFP_KERNEL for the allocation, and
both put the result of the allocation into the properties "mem"
structure.

Move the DMA allocation into gsi_alloc_evt_ring() instead.  Pass the
size to allocate as a new parameter.  The result will still be
placed into the properties "mem" structure, so it's available to the
caller (ipa_gsi_setup_channel() needs this).

Note that the memory must be allocated before it can be validated.

Change-Id: Iecd3f22c400a1e998c1dc381828ff4b039bf3b25
Signed-off-by: Alex Elder <elder@codeaurora.org>