msm: ipa: check desc_fifo_size at build time
We check that the size passed to gsi_alloc_evt_ring() is a multiple
of the ring element size, and return an error if it is not.  The
ring element size is always set to a value known at compile time, so
we can do this check then instead.  Move the definitions of the
channel and event ring element sizes into "gsi.h" so they can be
used outside "gsi.c".

Change-Id: Id7f44d4847df5fa2006f17d6389463208eefbb35
Signed-off-by: Alex Elder <elder@codeaurora.org>