msm: gsi: do_div() produces a remainder
In gsi_validate_evt_ring_mem() and gsi_validate_channel_props(),
do_div() is used on a DMA address, then the result multiplied
again by the divisor to see if it matches the original address.
It will not if there's a remainder from the division.

In addition to computing the quotient, do_div() returns the
remainder.  So rather than re-multiplying, just check to see if a
remainder was produced.

Change-Id: Icb216c0cb1baaa6d50695b1fe0389b56d628aef1
Signed-off-by: Alex Elder <elder@codeaurora.org>