msm: gsi: drop "op" local variables
Get rid of the local "op" variable in several functions,
and use its value in the one place in each where it's needed.

Also modify a gsi_writel() call at the end of in gsi_alloc_evt_ring()
to use "val" as its first parameter like all the previous calls.

Change-Id: I9980def6ef0ecdce73c252aef49c74a47bddf25f
Signed-off-by: Alex Elder <elder@codeaurora.org>