msm: gsi: free ring memory in gsi_dealloc_evt_ring()
Have gsi_dealloc_evt_ring() match its allocation counterpart by
freeing the DMA memory allocated there.  Zero out the exclusive
and int_modt fields as well for good measure.

There are two callers of gsi_dealloc_evt_ring(), and each is
paired with a call to free the allocated DMA memory.  One is in
the error path for ipa_gsi_setup_channel(), and the other is
in ipa3_teardown_sys_pipe().  They both use ep->gsi_evt_ring_mem to
describe the memory to be freed; that got assigned immediately after
the event ring was successfully allocated to equal what was saved
in the event context.

Change-Id: I3f0b085ae5a2c5866e4b2457726f749df68fb590
Signed-off-by: Alex Elder <elder@codeaurora.org>