msm: ipa_dp: free gsi_evt_ring_mem in ipa_gsi_setup_channel()
In the error path for ipa_gsi_setup_channel() the DMA memory
allocated for the event ring is freed.  The call passes the
of the memory buffer found in the properties structure.

There is another field, ep->gsi_evt_ring_mem, which is made to
contain the same information, and once assigned it should be
considered its "owner".  So free the DMA memory using that
rather than what was found in the properties structure.

Switch to using struct assignment for ep->gsi_evt_ring_mem.

Change-Id: I1962c79d77cd1b8f810c2c46146c960175cae224
Signed-off-by: Alex Elder <elder@codeaurora.org>