msm: ipa_dp: kill ipa3_ep_context->gsi_evt_ring_mem
The only purpose for the gsi_evt_ring_mem field of the endpoint
context was to be able to free the allocated DMA memory associated
with the endpoint.  Now that's handled by gsi_dealloc_evt_ring() so
we don't need to keep a copy of the memory information any more.

Change-Id: I98b3a3af8a47cd163ee9b2930197025a66c3cbd3
Signed-off-by: Alex Elder <elder@codeaurora.org>