msm: ipa_dp: kill struct gsi_evt_ring_props
The only real user of the gsi_evt_ring_props structure was
gsi_alloc_evt_ring(), and it no longer requires it.  So remove
the "props" parameter to that function.  The two callers therefore
no longer need their local variables of that type, so delete those.
The ipa_request_gsi_channel_params structure (which is unused in the
IPA lite code) contained a field of that type and it can go away.
That leaves no more reference to the gsi_evt_ring_props structure,
so it can be removed.

Change-Id: Iec19aff2c6799865e39e7a28f05690bd818c23c3
Signed-off-by: Alex Elder <elder@codeaurora.org>