msm: gsi: use barriers rather than mutex for clear_bit()
The event id used in gsi_alloc_evt_ring() is allocated under
protection of a mutex.  Once the corresponding bit is set in the
event mask, no other processor will use that event id.

Since clear_bit() is an atomic operation, we don't really need to
lock others out of access to the mask.  However we should use memory
barriers around--before it so no other processor starts using the
corresponding slot in the endpoint array, and (less importantly)
afterward to ensure the id is immediately available for allocation.

Change-Id: If38aa7644120f331558c3f83ffaf591ce95c9159
Signed-off-by: Alex Elder <elder@codeaurora.org>