msm: gsi: kill gsi_chan_props->max_prefetch
There are two places where the max_prefetch property of a channel
are set, and they both use GSI_ONE_PREFETCH_SEG (1) as the value.
Just assume that, and remove the "max_prefetch" field from the
channel properties struucture.

Define GSI_MAX_PREFETCH (= 1) in "gsi.c" and use that as the value
when programming the hardware.  Remove the definition of the
gsi_max_prefetch enumerated type.

Change-Id: I700460301415bb78a27aa483b9dd7cf7bab6ee6b
Signed-off-by: Alex Elder <elder@codeaurora.org>