msm: ipa: define IPA_HW_MIN
Define IPA_HW_MIN to be the minimum IPA hardware version supported
by the code.  Add comments to that and IPA_HW_MIN to explain their
(slightly different) interpretation.

Add a check in ipa_version_get() to report the version number and
return IPA_HW_None if the version found in the DTB is out of the
range supported by the code.  Have that function enforce the
"only version 3.5.1" rule, generically.

Delete a pointless redefinition of IPA_HW_v4_0.

Change-Id: Id05a9c2e0453511fa57af48735da072cffd66531
Signed-off-by: Alex Elder <elder@codeaurora.org>