msm: ipa: redefine value of IPA_HW_None
IPA_HW_None is used only to indicate an error returned by
ipa_version_get().  Other (valid) version numbers are only used in
the initialization routines and never thereafter.  Nothing assumes
IPA_HW_None has a particular value (and specifically, nothing
assumes it's 0).

Change the value used for IPA_HW_None to be the maximum possibly
represented in 32 bits (because the version is held in a 32-bit
value in the DTB).  This is preparing for an upcoming next patch,
which allows 0 to be a valid version identifier.

Change-Id: I3858024224d42269e1c8e6303bc0db51f18296e5
Signed-off-by: Alex Elder <elder@codeaurora.org>