msm: ipa: kill IPA_HW_MAX
Nothing needs to know the number of supported hardware versions.  If
a version is supported, it needs to have that support implemented
explicitly in ipa_version_get().  In addition, any supported version
will need to have an explicit entry in each of the IPA HAL-like arrays
(even if it inherits everything from an earlier version).

Get rid of IPA_HW_MAX, which is unused.

Change-Id: I6dbb3232fa7ca7ef3fa3895fe736f042f40cac2d
Signed-off-by: Alex Elder <elder@codeaurora.org>