msm: ipa: kill gsi_ctx->ipc_logbuf
The ipc_logbuf field of the gsi_ctx structure is never used, so get
rid of it.  That means there's no need for the "logbuf" parameter in
msm_gsi_init(), so get rid of that too.  (This should have happened
when we switched over to the GSI code using ipa_err(), etc.)

Change-Id: Ic0cb321d9c7a29d507f978f9fa45d852a0148bc1
Signed-off-by: Alex Elder <elder@codeaurora.org>