msm: gsi: limit event ids in gsi_alloc_evt_ring()
In gsi_alloc_evt_ring(), the event bitmap in the GSI context is
scanned for an available event id to use.  The scan considers up to
BITS_PER_LONG events (as many as could be represented by the mask),
but the maximum supported by the code is GSI_EVT_RING_MAX (= 23).

Impose the limit at the point we're allocating an event id, rather
than having to deal with out-of-range values later.

Change-Id: Ic9008d1c5a55dd05cd7742939c63a39b6693ceab
Signed-off-by: Alex Elder <elder@codeaurora.org>