msm: gsi: kill GSI_INST_RAM_n_WORD_SZ
Only one of the GSI_*_n_OFFS() macros uses a symbolic constant to
indicate the value the "n" parameter is multipled by to get the
offset.  Get rid of the definition of GSI_INST_RAM_n_WORD_SZ and
just use its value, to make the definition of GSI_INST_RAM_n_OFFS()
follow the same pattern as all the rest.

Change-Id: I141e95e51512791e5263326ac28fc6cbd8a2491c
Signed-off-by: Alex Elder <elder@codeaurora.org>