msm: gsi: change some constant names for uniqueness
The next patch will be shortening the names of all bitmask and shift
values defined in "gsi_reg.h", so they don't incorporate the name of
the register they affect.  In most cases the resulting shortened
name is still unique.  But some are not.  This patch changes some of
these names so that when the register name prefixes are stripped off
the resulting bitmask and shift symbol names remain unique.

Change-Id: I7f39bc491b70ac5af286cf9c775cbd40e4c5fe1d
Signed-off-by: Alex Elder <elder@codeaurora.org>