msm: gsi: format *_BMSK symbol definitions consistently
Use hexadecimal for all *_BMSK and $_OFST values.  Use 8 hex digits
for each base address (0x12345678).  For *_OFST(n) macros, use 4 hex
digits for the multipler, and for *_OFST(k, n) use 4 hex digits for
the first multiplier and 2 hex digits for the second.  For *_OFST(n)
and *_OFST(k, n) macros, the *_MAXn value is decimal, and for the
latter the *_MAXk value is also decimal.  Finally, there are a
couple field value symbols (*_FVAL) and they're decimal.

Change-Id: I4efb32121ebf1825e3c5b4499c8edb48d0eeb7cd
Signed-off-by: Alex Elder <elder@codeaurora.org>