msm: ipa_utils: have ALIGN_CHECK() return a value
Have ALIGN_CHECK() return the value of the symbol name it's passed,
in addition to verifying it meets its alignment requirements.  This
allows both things to occur in a single statement, reducing code and
avoiding a class of problems that arise from having to do them
separately.  Move the #undef of this and other macros past the end
of mem_partition_valid().

Change-Id: Idc42b13432617c79514bbd930eda30a5d65d7402
Signed-off-by: Alex Elder <elder@codeaurora.org>