msm: gsi: rename some parameters
Rename the "val" parameter to __gsi_write_channel_scratch() as well
as gsi_write_channel_scratch() to be "scr" rather than "val".  This
matches usage in gsi_get_channel_cfg() and gsi_set_channel_cfg(),
and is a better name than "val" (which is typically used for holding
a register value).

Change-Id: Id0226f2412bedabb83b3697b1c230c35bc4c85f4
Signed-off-by: Alex Elder <elder@codeaurora.org>