msm: gsi: check for excess channel/event bits outside loop
Several functions read a register containing a bitmask representing
channels or events that need some sort of processing.  Within the
loop there is a check to see if the loop index has gone past the
maximum.

Rather than doing this, check the mask value right after it is read
to see if any bits are set that represent something more than
expected.  Then have the loop iterate only over in-range values.
We know that the max_ch and max_ev fields of the GSI context are
non-zero (gsi_register_device() would have failed if not).

Note that we'll make this check in two functions that didn't do this
before: gsi_handle_inter_ee_ch_ctrl() and gsi_handle_inter_ee_ev_ctrl().

Change-Id: Ie6ddaa5cd2e688810d8abff72ced86e7299a1986
Signed-off-by: Alex Elder <elder@codeaurora.org>