msm: ipa_dp: shift out another block in ipa3_assign_policy()
The previous patch explained that only four client values were ever
passed to ipa3_assign_policy().  Two are handled early in that
function, and the other two are covered by a large conditional block
at the end of that function.  Because we know what values "client" can
hold, we know that block can be made unconditional (and we can
remove an error case as well).

Change-Id: I4e6394dfe3c27f5f073d4326c9c16da277699a55
Signed-off-by: Alex Elder <elder@codeaurora.org>