msm: gsi: chan_hdl is good in gsi_write_channel_scratch()
Only ipa_gsi_setup_channel() calls gsi_write_channel_scratch(),
and the "chan_hdl" value it supplies is the result of a successful
call to gsi_alloc_channel().  The handle value returned by that
function is the same as the "ch_id" the caller supplies in its
channel properties structure.  For to gsi_alloc_channel() to
return successfully, the properties must have been validated by
gsi_validate_channel_props(), and that function returns an error
if the "ch_id" field is too large.

So we know that the "chan_hdl" that gsi_write_channel_scratch()
recieves is in range, and we don't have to check that condition.

Change-Id: Ibffb262911133148486652b7593078ad69fc4b9a
Signed-off-by: Alex Elder <elder@codeaurora.org>