msm: gsi: make ipa_gsi_ep_config->ipa_if_aos and ->ee unsigned
The "ipa_if_aos" field of ipa_gsi_ep_config structure is only ever
assigned by the IPA lite code, not used.  None of the values
assigned to that field is negative, so there's no harm in changing
its type to be unsigned.  Change the "ee" field of that structure to
be unsigned as well.  All EE values are small non-negative numbers.

Change-Id: I6167e7ede63674d6fe2ab661081f4dea55f690eb
Signed-off-by: Alex Elder <elder@codeaurora.org>