msm: ipa_dp: fix cleanup path in ipa_gsi_setup_channel()
At the end of ipa_gsi_setup_channel(), there is a conditional
deallocation of an event ring.  This should happen only if we're not
using the common event ring, but currently it does so if the handle
is not GSI_NO_EVT_ERINDEX.  This is a bug.

Change the condition so it deallocates the event ring only when one
was allocated--that is, when we were not going to use the common one.

Change-Id: I8c07a81ee5899438bd00ff44e6d7321e35ac62a9
Signed-off-by: Alex Elder <elder@codeaurora.org>