msm: gsi: init spinlock in gsi_init_ring()
The only field not initialized in gsi_init_ring() is the spinlock.
It was only being initialized when its corresponding channel or
event was allocated; now it will be initialized along with the
rest of the ring.  Reinitializing the spinlock may technically not
be necessary, but logically it's the right thing to do.

Change-Id: I8bb6bf52708d7073901b8661295ccffddf473cec
Signed-off-by: Alex Elder <elder@codeaurora.org>