msm: gsi: verify in-range channel number in ipa_gsi_setup_channel()
The maximum number of channels supported is determine at runtime.
Meanwhile, GSI channel numbers are defined by entries in the
ipa3_ep_configuration[][] array.  We first look an endpoint's
channel number in ipa_gsi_setup_channel(), so that's a good
opportunity to verify the value from the array is within the
supported range.  Add that check so we can avoid doing so later
following successful channel allocation.  Do so before getting the
ring handle, because it can be done without acquiring any resources.

Change-Id: I0bb5de67c21bac8a4c73fe722c673b5e7a891776
Signed-off-by: Alex Elder <elder@codeaurora.org>