msm: gsi: no need for spinlock pointers
There are two functions that hold the address of a spinlock in
local variables.  The local variable was needed at one time to refer
to one of two possible lock structures, but there is no need for
that any more.  Get rid of the local variables and just pass the
address of the spinlock directly to the locking functions.

Change-Id: Ie240d5ed60d12576404af926458be4f7a0a9a5f5
Signed-off-by: Alex Elder <elder@codeaurora.org>