msm: gsi: simplify gsi_config_channel_mode()
There are only two channel modes: CALLBACK and POLL.  A channel
context's poll_mode set initially to CALLBACK in gsi_alloc_channel(),
and thereafter it is only updated in gsi_config_channel_mode().
There are only three cases in this function:  the mode supplied is
the same as what's currently in effect; the mode is POLL and we're
currently in CALLBACK mode; or the mode is CALLBACK and we're
currently in POLL mode.  We can simplify the logic based on this
knowlege.

A few assertion calls are put in place but commented out until I
make some changes to how ipa_assert() works (sometime soon).

Change-Id: I5bf9e80fd5b605e6cd82a595d6a8e16fd4dc1333
Signed-off-by: Alex Elder <elder@codeaurora.org>