msm: gsi: don't use devm_kzalloc() for channel user data
A channel context has block of pointers allocated for "user_data"
when the channel is allocated.  It's allocated using devm_kzalloc().
We're allocating this outside the normal probe and teardown of the
device, and we're managing the life cycle of this allocated block
OK, so just use the normal kzalloc()/kfree() pair rather than
incurring the overhead for the device managed variant.

Change-Id: I13ed7946ae0c9068a90856f8ad5e7f26e006b4d3
Signed-off-by: Alex Elder <elder@codeaurora.org>