msm: gsi: kill max_elem_dp_stats debugfs file
Remove support for the /sys/kernel/debug/gsi/max_elem_dp_stats
file in the debug file system.  It was only used to update the
"max_re_expected" field of a channel properties structure, and that
was only used in maintaining the channel data path statistics
removed by the previous patch.

This of course means we should remove the "max_re_expected" field
as well; it is assigned but never used.

Change-Id: I070c01facc1c0990d06044b491eeba69fae6c23e
Signed-off-by: Alex Elder <elder@codeaurora.org>