msm: gsi: defer last channel ring update in gsi_process_chan()
In gsi_process_chan() the channel id is extracted from an event,
and the transfer pointer in the event is used to determine which
channel ring element has completed.  Both read pointers for the ring
are updated to indicate the channel entry is complete, and a
transfer callback function is (optionally) called for that completed
ring element.

Currently the read pointers are updated before calling the transfer
callback.  This patch defers doing that until afterward.

Change-Id: I6c6f03a6a56563920014ccb9ab1ec014c3280bab
Signed-off-by: Alex Elder <elder@codeaurora.org>