msm: gsi: have gsi_process_evt_re() return transfer count
Have gsi_process_evt_re() return the number of bytes that the event
ring element indicates got transferred.  This is the value that is
placed into notify->bytes_xfered.  Move the declaration of the
notify structures passed to this function to the same block that
calls it; this makes it obvious the structure is otherwise unused.

Change-Id: I172cf43eaa4b1c89b30d7ece2ba8add159f6c514
Signed-off-by: Alex Elder <elder@codeaurora.org>