msm: gsi: have gsi_process_chan() return transfer count
Have gsi_process_chan() return the number of bytes that the
completion event indicates got transferred.  This is the same
as the value placed into notify->bytes_xfered.

This value is ignored if "callback" is true--gsi_handle_ieob()
ignores the result of gsi_process_evt_re(), which returns whatever
gsi_process_chan() returns.  The place it is used is in
gsi_poll_channel(), which returns what gsi_process_evt_re() returns,
which is the transfer count.

Change-Id: I23c478746015c0f6967cbfabbc8a3f4e0cf5abe6
Signed-off-by: Alex Elder <elder@codeaurora.org>