msm: ipa_dp: drop meta argument to ipa3_tx_dp()
The only caller of ipa3_tx_dp in IPA lite is ipa3_wwan_xmit(), and
that passes NULL as its "meta" argument.  Drop that argument and
adjust the code based on the knowledge it would be NULL.  This
removes the need for the ipa_tx_meta structure, so it too can be
removed.

Change-Id: Idfc150fd426d894e78932a7d4b016bcd0dc1f620
Signed-off-by: Alex Elder <elder@codeaurora.org>