msm: ipa_dp: kill ipa3_desc->is_tag_valid
The is_tag_valid flagl in the ipa3_desc structure is never set.  So
remove the field and delete a block of code in ipa3_send() that was
only executed if it was true.  This makes ipa_populate_tag_field()
an unused function, so get rid of that also.  And removing that
function makes pointer_to_tag_wa() no longer used, so it too can
go away.

Change-Id: I5e6ff2bbe8c63b0be2857429f2726c43a234c8af
Signed-off-by: Alex Elder <elder@codeaurora.org>