msm: ipa: construct payload in ipahal_dma_shared_mem_write_pyld()
Payload allocation functions fill in a parameter structure, then
call ipahal_construct_imm_cmd() to allocate and fill a separate
structure that that will get sent to the hardware.  Rather than
doing that we can just skip the middleman fill the hardware
structure directly.  This patch is the first in a series that
converts code to use this model.

Here, we allocate a ipa_imm_cmd_hw_dma_shared_mem structure
in ipahal_dma_shared_mem_write_pyld(), then fill it in directly in
the same way that ipa_imm_cmd_construct_dma_shared_mem() would.

Change-Id: I4edb7b7934ce6c6b55b01e9c5a9a261ea2733554
Signed-off-by: Alex Elder <elder@codeaurora.org>