msm: ipa: kill struct ipa_imm_cmd_hw_ip_v6_routing_init
The hardware structures for initializing IP routing tables are for
IPv4 and IPv6 are identical.  Get rid of the IP version-specific
structures and just use a common one.  This makes it obvious that
they are the same.  If we find they differ at some future point we
will define separate ones.

Change-Id: I65352032f2de44898be0c3459938494d61eef050
Signed-off-by: Alex Elder <elder@codeaurora.org>